Commit api and db changes and for results report and survey stats and overview change for branch

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/SurveyStatsReportBranch.java
@@ -74,7 +74,7 @@ public class SurveyStatsReportBranch
     private long partiallyCompleted;
     
     @Column( name = "complete_percentage")
-    private long completePercentage;
+    private double completePercentage;
     
     @Column( name = "delta")
     private long delta;
@@ -289,12 +289,12 @@ public class SurveyStatsReportBranch
         this.partiallyCompleted = partiallyCompleted;
     }
 
-    public long getCompletePercentage()
+    public double getCompletePercentage()
     {
         return completePercentage;
     }
 
-    public void setCompletePercentage( long completePercentage )
+    public void setCompletePercentage( double completePercentage )
     {
         this.completePercentage = completePercentage;
     }
